requirements.txt
python-dotenv
requests
pydantic
google-generativeai
openai
tavily-python
rich
edge-tts
pygame
